Support more 10_10_10_2 formats as vertex buffers
Vertex input buffers only support VK_FORMAT_A2B10G10R10_UNORM_PACK32.
This cl adds all the signed and integer variants of that format.

This cl also removes StreamType and replaces it with VkFormat.

Tests: dEQP-VK.pipeline.vertex_input.*r10*
Bug: b/142661203

+++ src/Vulkan/VkFormat.cpp
@@ -1103,7 +1103,7 @@ int Format::bytesPerBlock() const
 	}
 }
 
-int Format::componentCount() const
+uint8_t Format::componentCount() const
 {
 	switch(format)
 	{
@@ -1272,7 +1272,7 @@ int Format::componentCount() const
 	return 1;
 }
 
-bool Format::isUnsignedComponent(int component) const
+bool Format::isUnsignedComponent(uint8_t component) const
 {
 	switch(format)
 	{
@@ -2242,7 +2242,7 @@ bool Format::has32bitIntegerTextureComponents() const
 	return false;
 }
 
-bool Format::isRGBComponent(int component) const
+bool Format::isRGBComponent(uint8_t component) const
 {
 	switch(format)
 	{
